+++ package.json
@@ -1,9 +1,10 @@
 {
   "name": "blackdeserthorsetiming",
-  "version": "0.0.0",
+  "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "node ./app.js",
+	"build": "webpack"
   },
   "dependencies": {
     "body-parser": "~1.16.0",
@@ -14,5 +15,8 @@
     "jsonfile": "^2.4.0",
     "morgan": "~1.7.0",
     "serve-favicon": "~2.3.2"
+  },
+  "devDependencies": {
+    "webpack": "^2.2.1"
   }
 }
