+++ public/timers.html
@@ -7,10 +7,9 @@
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" integrity="sha384-1q8mTJOASx8j1Au+a5WDVnPi2lkFfwwEAa8hDDdjZlpLegxhjVME1fgjWPGmkzs7" crossorigin="anonymous">
     <link rel="stylesheet" href="/stylesheets/style.css">
 
-    <script src="/javascripts/timings.js"></script>
-    <script src="/javascripts/timingsvue.js"></script>
+    <script src="/javascripts/timings.bundle.js"></script>
 </head>
-<body onload="initVue()">
+<body onload="initAppz()">
 <div class="btn-danger" style="text-align:center; font-size: 150%;">Page only support viewing in Google Chrome at the
     moment
 </div>
